FIX: remove rendering UX from ember models
Fix for `admin-user.js` and `backup.js` models.
