+++ app/assets/javascripts/admin/addon/controllers/admin-user-index.js
@@ -1,16 +1,20 @@
+import DiscourseURL, { userPath } from "discourse/lib/url";
 import { and, notEmpty } from "@ember/object/computed";
 import { fmt, propertyNotEqual, setting } from "discourse/lib/computed";
+import AdminUser from "admin/models/admin-user";
 import CanCheckEmails from "discourse/mixins/can-check-emails";
 import Controller from "@ember/controller";
 import I18n from "I18n";
 import { ajax } from "discourse/lib/ajax";
 import bootbox from "bootbox";
 import discourseComputed from "discourse-common/utils/decorators";
+import getURL from "discourse-common/lib/get-url";
 import { htmlSafe } from "@ember/template";
+import { iconHTML } from "discourse-common/lib/icon-library";
+import messageBus from "message-bus-client";
 import { popupAjaxError } from "discourse/lib/ajax-error";
 import { inject as service } from "@ember/service";
 import showModal from "discourse/lib/show-modal";
-import { userPath } from "discourse/lib/url";
 
 export default Controller.extend(CanCheckEmails, {
   adminTools: service(),
@@ -141,10 +145,21 @@ export default Controller.extend(CanCheckEmails, {
 
   actions: {
     impersonate() {
-      return this.model.impersonate();
+      return this.model
+        .impersonate()
+        .then(() => (document.location = getURL("/")))
+        .catch((e) => {
+          if (e.status === 404) {
+            bootbox.alert(I18n.t("admin.impersonate.not_found"));
+          } else {
+            bootbox.alert(I18n.t("admin.impersonate.invalid"));
+          }
+        });
     },
     logOut() {
-      return this.model.logOut();
+      return this.model
+        .logOut()
+        .then(() => bootbox.alert(I18n.t("admin.user.logged_out")));
     },
     resetBounceScore() {
       return this.model.resetBounceScore();
@@ -153,19 +168,43 @@ export default Controller.extend(CanCheckEmails, {
       return this.model.approve(this.currentUser);
     },
     deactivate() {
-      return this.model.deactivate();
+      return this.model
+        .deactivate()
+        .then(() => window.location.reload())
+        .catch((e) => {
+          const error = I18n.t("admin.user.deactivate_failed", {
+            error: this._formatError(e),
+          });
+          bootbox.alert(error);
+        });
     },
     sendActivationEmail() {
-      return this.model.sendActivationEmail();
+      return this.model
+        .sendActivationEmail()
+        .then(() => bootbox.alert(I18n.t("admin.user.activation_email_sent")))
+        .catch(popupAjaxError);
     },
     activate() {
-      return this.model.activate();
+      return this.model
+        .activate()
+        .then(() => window.location.reload())
+        .catch((e) => {
+          const error = I18n.t("admin.user.activate_failed", {
+            error: this._formatError(e),
+          });
+          bootbox.alert(error);
+        });
     },
     revokeAdmin() {
       return this.model.revokeAdmin();
     },
     grantAdmin() {
-      return this.model.grantAdmin();
+      return this.model
+        .grantAdmin()
+        .then(() => {
+          bootbox.alert(I18n.t("admin.user.grant_admin_confirm"));
+        })
+        .catch(popupAjaxError);
     },
     revokeModeration() {
       return this.model.revokeModeration();
@@ -174,13 +213,41 @@ export default Controller.extend(CanCheckEmails, {
       return this.model.grantModeration();
     },
     saveTrustLevel() {
-      return this.model.saveTrustLevel();
+      return this.model
+        .saveTrustLevel()
+        .then(() => window.location.reload())
+        .catch((e) => {
+          let error;
+          if (e.jqXHR.responseJSON && e.jqXHR.responseJSON.errors) {
+            error = e.jqXHR.responseJSON.errors[0];
+          }
+          error =
+            error ||
+            I18n.t("admin.user.trust_level_change_failed", {
+              error: this._formatError(e),
+            });
+          bootbox.alert(error);
+        });
     },
     restoreTrustLevel() {
       return this.model.restoreTrustLevel();
     },
     lockTrustLevel(locked) {
-      return this.model.lockTrustLevel(locked);
+      return this.model
+        .lockTrustLevel(locked)
+        .then(() => window.location.reload())
+        .catch((e) => {
+          let error;
+          if (e.jqXHR.responseJSON && e.jqXHR.responseJSON.errors) {
+            error = e.jqXHR.responseJSON.errors[0];
+          }
+          error =
+            error ||
+            I18n.t("admin.user.trust_level_change_failed", {
+              error: this._formatError(e),
+            });
+          bootbox.alert(error);
+        });
     },
     unsilence() {
       return this.model.unsilence();
@@ -189,11 +256,121 @@ export default Controller.extend(CanCheckEmails, {
       return this.model.silence();
     },
     deleteAllPosts() {
-      return this.model.deleteAllPosts();
+      let deletedPosts = 0;
+      const user = this.model;
+      const message = I18n.messageFormat(
+        "admin.user.delete_all_posts_confirm_MF",
+        {
+          POSTS: user.get("post_count"),
+          TOPICS: user.get("topic_count"),
+        }
+      );
+
+      const performDelete = () => {
+        let deletedPercentage = 0;
+        this.model
+          .deleteAllPosts()
+          .then(({ posts_deleted }) => {
+            if (posts_deleted === 0) {
+              user.set("post_count", 0);
+              bootbox.hideAll();
+            } else {
+              deletedPosts += posts_deleted;
+              deletedPercentage = Math.floor(
+                (deletedPosts * 100) / user.get("post_count")
+              );
+              $(".delete-posts-progress .progress-bar > span").css({
+                width: `${deletedPercentage}%`,
+              });
+              performDelete();
+            }
+          })
+          .catch((e) => {
+            bootbox.hideAll();
+            let error;
+            AdminUser.find(user.get("id")).then((u) => user.setProperties(u));
+            if (e.jqXHR.responseJSON && e.jqXHR.responseJSON.errors) {
+              error = e.jqXHR.responseJSON.errors[0];
+            }
+            error = error || I18n.t("admin.user.delete_posts_failed");
+            bootbox.alert(error);
+          });
+      };
+
+      const buttons = [
+        {
+          label: I18n.t("composer.cancel"),
+          class: "d-modal-cancel",
+          link: true,
+        },
+        {
+          label:
+            `${iconHTML("exclamation-triangle")} ` +
+            I18n.t("admin.user.delete_all_posts"),
+          class: "btn btn-danger",
+          callback: () => {
+            openProgressModal();
+            performDelete();
+          },
+        },
+      ];
+      const openProgressModal = () => {
+        bootbox.dialog(
+          `<p>${I18n.t(
+            "admin.user.delete_posts_progress"
+          )}</p><div class='progress-bar'><span></span></div>`,
+          [],
+          { classes: "delete-posts-progress" }
+        );
+      };
+      bootbox.dialog(message, buttons, { classes: "delete-all-posts" });
     },
+
     anonymize() {
-      return this.model.anonymize();
+      const user = this.model;
+      const message = I18n.t("admin.user.anonymize_confirm");
+
+      const performAnonymize = () => {
+        this.model
+          .anonymize()
+          .then((data) => {
+            if (data.success) {
+              if (data.username) {
+                document.location = getURL(
+                  `/admin/users/${user.get("id")}/${data.username}`
+                );
+              } else {
+                document.location = getURL("/admin/users/list/active");
+              }
+            } else {
+              bootbox.alert(I18n.t("admin.user.anonymize_failed"));
+              if (data.user) {
+                user.setProperties(data.user);
+              }
+            }
+          })
+          .catch(() => bootbox.alert(I18n.t("admin.user.anonymize_failed")));
+      };
+      const buttons = [
+        {
+          label: I18n.t("composer.cancel"),
+          class: "cancel",
+          link: true,
+        },
+        {
+          label:
+            `${iconHTML("exclamation-triangle")} ` +
+            I18n.t("admin.user.anonymize_yes"),
+          class: "btn btn-danger",
+          callback: () => {
+            performAnonymize();
+          },
+        },
+      ];
+
+      bootbox.dialog(message, buttons, { classes: "delete-user-modal" });
     },
+
     disableSecondFactor() {
       return this.model.disableSecondFactor();
     },
@@ -210,11 +387,69 @@ export default Controller.extend(CanCheckEmails, {
     destroy() {
       const postCount = this.get("model.post_count");
       const maxPostCount = this.siteSettings.delete_all_posts_max;
-      if (postCount <= maxPostCount) {
-        return this.model.destroy({ deletePosts: true });
-      } else {
-        return this.model.destroy();
-      }
+
+      const user = this.model;
+      const message = I18n.t("admin.user.delete_confirm");
+      const location = document.location.pathname;
+
+      const performDestroy = (block) => {
+        bootbox.dialog(I18n.t("admin.user.deleting_user"));
+        let formData = { context: location };
+        if (block) {
+          formData["block_email"] = true;
+          formData["block_urls"] = true;
+          formData["block_ip"] = true;
+        }
+        if (postCount <= maxPostCount) {
+          formData["delete_posts"] = true;
+        }
+        this.model
+          .destroy(formData)
+          .then((data) => {
+            if (data.deleted) {
+              if (/^\/admin\/users\/list\//.test(location)) {
+                document.location = location;
+              } else {
+                document.location = getURL("/admin/users/list/active");
+              }
+            } else {
+              bootbox.alert(I18n.t("admin.user.delete_failed"));
+              if (data.user) {
+                user.setProperties(data.user);
+              }
+            }
+          })
+          .catch(() => {
+            AdminUser.find(user.get("id")).then((u) => user.setProperties(u));
+            bootbox.alert(I18n.t("admin.user.delete_failed"));
+          });
+      };
+
+      const buttons = [
+        {
+          label: I18n.t("composer.cancel"),
+          class: "btn",
+          link: true,
+        },
+        {
+          label:
+            `${iconHTML("exclamation-triangle")} ` +
+            I18n.t("admin.user.delete_and_block"),
+          class: "btn btn-danger",
+          callback: () => {
+            performDestroy(true);
+          },
+        },
+        {
+          label: I18n.t("admin.user.delete_dont_block"),
+          class: "btn btn-primary",
+          callback: () => {
+            performDestroy(false);
+          },
+        },
+      ];
+
+      bootbox.dialog(message, buttons, { classes: "delete-user-modal" });
     },
 
     promptTargetUser() {
@@ -235,7 +470,43 @@ export default Controller.extend(CanCheckEmails, {
     },
 
     merge(targetUsername) {
-      return this.model.merge({ targetUsername });
+      const user = this.model;
+      const location = document.location.pathname;
+
+      const bootboxDiv = bootbox.dialog(I18n.t("admin.user.merging_user"));
+      let formData = { context: location };
+
+      if (targetUsername) {
+        formData["target_username"] = targetUsername;
+      }
+
+      this.model
+        .merge(formData)
+        .then((response) => {
+          if (response.success) {
+            messageBus.subscribe("/merge_user", (data) => {
+              if (data.merged) {
+                if (/^\/admin\/users\/list\//.test(location)) {
+                  DiscourseURL.redirectTo(location);
+                } else {
+                  DiscourseURL.redirectTo(
+                    `/admin/users/${data.user.id}/${data.user.username}`
+                  );
+                }
+              } else if (data.message) {
+                bootboxDiv.find(".modal-body").html(data.message);
+              } else if (data.failed) {
+                bootbox.alert(I18n.t("admin.user.merge_failed"));
+              }
+            });
+          } else {
+            bootbox.alert(I18n.t("admin.user.merge_failed"));
+          }
+        })
+        .catch(() => {
+          AdminUser.find(user.id).then((u) => user.setProperties(u));
+          bootbox.alert(I18n.t("admin.user.merge_failed"));
+        });
     },
 
     viewActionLogs() {
